Add files via upload
The Higgs Boson Challenge on Kaggle, also known as the "Higgs Boson Machine Learning Challenge," was a competition hosted on Kaggle in collaboration with CERN (the European Organization for Nuclear Research). The challenge aimed to encourage data scientists and machine learning practitioners to develop models for the classification of Higgs boson events in high-energy particle physics.